pyproject.toml
[tool.poetry]
name = "image_morphing"
version = "0.1.0"
description = ""
authors = ["fasfa <afsafa>"]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.5.2"
dlib = "^19.22.0"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
